Fix some compiler warnings
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/3500



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@15468 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/test/intltest/icusvtst.cpp
@@ -153,7 +153,7 @@ UnicodeString append(UnicodeString& result, const UObject* obj)
         } else if (id == Locale::getStaticClassID()) {
             result.append(((Locale*)obj)->getName());
         } else if (id == Integer::getStaticClassID()) {
-            sprintf(buffer, "%d", ((Integer*)obj)->_val);
+            sprintf(buffer, "%d", (int)((Integer*)obj)->_val);
             result.append(buffer);
         } else {
             sprintf(buffer, "%p", (const void*)obj);
